requirements.txt
aniso8601==7.0.0
pwned-passwords-django==1.4
django-axes==5.6.1
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.9
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-taggit==1.3.0
html5lib==1.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
promise==2.3
psycopg2-binary==2.8.5
pyaes==1.6.1
pycodestyle==2.6.0
PyJWT==1.7.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
Rx==1.6.1
sentry-sdk==0.17.3
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
TgCrypto==1.2.1
toml==0.10.1
tzlocal==2.1
Unidecode==1.1.1
django-user-visit==0.4
urllib3==1.25.8
webencodings==0.5.1
